Trim HTML in replies to the contents of the <body> tag
Email clients vary as to if they attempt to provide a full HTML document
(starting with doctype and <html>), or merely an HTML-formatted snippet.
Content which contains an <html> tag unfortunately interacts poorly with
CKEditor when shoved inside a <blockquote>, effectively imediately
closing the blockquote.

Pass content to quote through an HTML scrubber to remove the offending
<html> tags, the <head> tag, and and any <meta> contents thereof.