Update test; whitespace is allowable, and does not cause an update
This also steps around a bug in WWW::Mechanize -- namely, while real
browsers ignore a newline that directly follows a <textarea>,
WWW::Mechanize does not.  As such, it views <textarea>\n</textarea> as
containing a newline, whereas browsers view it as empty.

In this case, bulk update ignores it unless it has non-whitespace,
causing the difference to be immaterial.