Ensure that panes are displayed in a stable order
'keys' on a hash is no longer stable in recent versions of Perl.  As
such, the display order of the panes (as pushed onto @panes at the end
of the block) was no longer stable.  This caused the Sidebar and Body
sections to randomly flip back and forth as a portlet was moved up and
down in them, or other edits were made.

Sort the panes by name, to ensure consistent display order.