_EncodeLOB expects to get an octet string.
We were passing the character string returned from _DecodeLOB and
persisted by Storable.  This caused checks of validity to fail and
things that should have been encoded in Quoted-Printable to instead end
up being passed directly to PostgreSQL which kicks them back as invalid
UTF-8.

Combined with our change to _DecodeLOB to now strictly decode to UTF-8
and apply Encode::PerlQQ, we should no longer be able to pass invalid
data along to the database.  The Encode::FB_Croak enforces this before
we get to the DB.