Allow searching for dates which are unset
e60885aa changed the meaning of "Starts < 'today'" (and all other date
limits) to no longer include tickets whose dates are unset.  This
behavior change also made it imposible to search for such tickets, as
"Starts = '1970-01-01'" would also no longer find them, and the more
straightforward "Starts IS NULL" was invalid.

Allow IS NULL and IS NOT NULL searches on date ranges.