Pass Heavy => 1 to InitClasses to enable error checking
8976c535 erroneously claimed that it was replacing a clone of RT->Init,
but missed that RT->Init doesn't call InitClasses(Heavy => 1) but
standalone_httpd did.

When standalone_httpd and rt-server were mashed together during the
PSGIfication, Heavy => 1 had already been removed. Sadly - the other
Heavy => 1 usage in webmux.pl was lost in e1c6d4a8c.

This means that no version of RT 4.0.x or 4.2.x has checked if your
Scrips had a valid Condition and Action, or checked your classes in
CustomFieldValuesSources.  It also misses the pre-loading we were trying
to do for the sake of mod_perl (21f181dd).

It's likely the CustomFieldValuesSources check could go in a
PostLoadCheck, but the Scrips code would need to keep living here.
The LoadLexicons which was also added for caching (21f181dd) was removed
a month before all the Heavy code broke in order to improve test
performance 5f2799c5, it's unclear if it should come back in some cases.