Allow queue CF rights to apply on single-queue searches
Because of the lack of context object, granting a user SeeCustomField on
a queue was insufficient to cause its CFs to appear if the search was
limited to that queue.  While CFs cannot currently have more than one
context object (as would be necessary in the case of a search over
multiple queues), the usability can be improved in the common case of a
limit on a single queue.

Apply the relevant queue as context object to CF limits when searching
within a single queue.  This allows "SeeCustomField" on the queue level
to be respected in the SearchBuilder.