Use HTML::FormatExternal to allow external HTML → text formatters
The HTML::FormatText::WithLinks::AndTables module is unmaintained, and
contains multiple bugs which trigger runtime errors.  In such cases, RT
sends out an empty text/plain part -- if HTML templates are not in use,
this results in an entirely empty email.

Add HTML::FormatExternal as an optional dependency for more reliable
HTML to text conversion using external text-based web browsers and text
conversion tools.

Benchmarking shows that shelling out to an external program on complex
HTML is comparable in speed (or faster!) than the pure-perl formatter
(20/s).  On simple HTML, the pure-perl formatter is an order of
magnitude faster -- but the slowest external formatter (elinks) still
runs at 5/sec, and most run at 20-50/s.  This is judged to not be likely
to cause a performance impact, as it is run at most once per scrip.