Require newer Email::Address::List to prevent denial of service
Email::Address::List contained a regular expression which was
susceptible to a quadratic-time increase in runtime on some
incorrectly-formed strings; this allowed constructed emails to subject
RT to a denial-of-service attack.  Require a newer version of the
module, which alters the regular expression to no longer contain such a
pessimized case.

This resolves CVE-2014-1474.