Skip custom fields that do not apply to the ticket/transaction
RT::Interface::Web::CreateTicket munges the query parameters to pass
Customfield-42 to RT::Ticket::Create, as of be3529af.  An important
change made in this commit is that previously, transaction custom fields
were passed through with their keys unmodified -- that is, as
Object-RT::Transaction--Customfield-42.

After be3529af, the information about if the CF was a Ticket or
Transaction custom field has been lost by the time the CFs are being
iterated.  This led to attempting to add values for Transaction CFs on
Ticket objects, and vice versa -- and thus warnings in the UI of "Custom
field 42 does not apply to this object" for every applied transaction
CF.

Only attempt to add CF values to objects if they are of the correct
class.