Ensure that the query is re-run after a bulk update
While the CollectionList at the top of the Bulk Update re-runs the query
after an update (by dint of being passed the TicketSQL and not
$Tickets), the BulkLinks collection is not so lucky.  As such, it
contains the data from before the update was applied.

However, this can lead to internal inconsistency -- namely, if the
tickets were deleted, Count on $Tickets will return the old cached
value, while ->Next will skip the newly-deleted tickets.  This causes
BulkLinks to error out with "Can't call method "DependsOn" on an
undefined value", as it assumes ->Next will be true if ->Count was.

Ensure that the $Tickets collection is re-run if a bulk update was made.