Remove duplicate ParseCcAddressesFromHead in RT::EmailParser
This code was originally in RT::Interface::Email; in e6b3fba, it was
moved to RT::EmailParser.  However, in a90721aa it was added _back_ to
RT::Interface::Email -- presumably because the "new mailgate" was
derived from RT::Interface::Email just before bcb5c3ff, not the most
current version.

This mistake was slow to be undone -- a small bit was first done in
e48d70a8.  Then dad97a8e attempted to address it on 3.2-trunk, but was
reverted in f706a50a soas to not break the API during a stable series.
Unfortunately, after it was re-applied to 3.3-trunk in 6c2d22842, SVK
incorrectly put them back in the merge-up from 3.2-trunk in a6577c51.

4fad7e78 got most of them, but missed ParseCcAddressesFromHead because
of the refactorings in 19ac5467 had caused them to diverge.

There are no current call sites of the RT::EmailParser implementation of
ParseCcAddressesFromHead; remove it.  This removes the last method which
exists in both RT::EmailParser and RT::Interface::Email.