Issues #28636: Articles setup warning appearance
The issue was introduced in 61e7290e when the theme was created
which seems to be an attempt to style the display of simple error
msgs rendered by Elements/Error which is not in the titltebox
context.

The fix here was to make the pages using a titlebox to display an
error msg to use a different error class than just "error", that is,
now called "error-titlebox".