Remove CreateUser, merging to form a more featureful LoadOrCreateByEmail
CreateUser is a near-duplicate of LoadOrCreateByEmail -- which is not
surprising, since they share history.  The code originally arose in the
RT::Interface::Email in f51a612f, and was copied into the "add a
watcher" code in Tickets and Queues in 29fbc05f and 40b1ea0,
respectively.  The code in Ticket became LoadOrCreateByEmail in
bd37466e, but the code in Queue remained there until 97a6962a.

This removes the last duplication of this code, merging the additional
features from CreateUser into LoadOrCreateByEmail.  This also allows
RT::Crypt::SMIME to create users with a more correct comment string than
"Autocreated when added as a watcher", as it previously did.