Don't search both Queue and Ticket CFs when passed a Queue.
In e2c21b55 (new in 4.2) we began calculating LookupType from the
ContextObj set in bb24a9f4 (back in 4.0).  This resulted in a LookupType
of RT::Queue if you said:

    $cf->LoadByName( Name => 'foo', Queue => 'General' );

because the ContextObj was set to General and then used to calculate a
LookupType.

Then we call LimitToQueue which force-applies a LookupType of
RT::Queue-RT::Ticket, meaning that we now search for:

    (main.LookupType IN ('RT::Queue') OR main.LookupType = 'RT::Queue-RT::Ticket')

This enforces the backwards compatible 'intended' change that

    $cf->LoadByName( Name => 'foo', Queue => 'General' );

searches for a ticket CF applied to the General queue and named foo.
Current 4.2 would also find a foo Queue CF applied to General.

This is seen best by code using RT::Queue->CustomField to load a
specific Custom Field applied to a Queue (but not globally).  This
commit also fixes those docs because they could confusingly lead you to
believe that $Queue->CustomField('foo') would get you the Queue custom
field named foo.  It has always retrieved the Ticket CF dating back to
the original CF implementation, but the docs were never updated when we
began allowing Custom Fields on Queues.

If you want a Queue's Foo CF you need to call

    $cf->LoadByName( Name => 'foo', Queue => 'General', LookupType => 'RT::Queue' );