Always create the user; this simplifies ACL checking greatly
This is a behavior change.  Previously, if nonexistant@example.com sent
mail to a queue which they did not have rights to create a ticket, no
user would be created.  Now, the user is fully created -- this allows
them to, for example, possibly pick up more rights than just Everyone
and Unprivileged -- before ticket creation is attempted.

This simplifies ACL checking greatly, as the GetCurrentUser code no
longer needs to wave its hands about Everyone and Unprivileged, and
attempt to check rights before it actually has a CurrentUser.  It
additionally allows users to be created with additional rights (pulled,
say, from LDAP).

In deployments which do not grant CreateTicket to Everyone or
Unprivileged, this will result in more unprivileged users.  However, the
combination of deployments that accept arbitrary mail from the outside
world, as well as not granting CreateTicket, is rather rare -- and in
fact, an email is sent to OwnerEmail every time those constraints were
satisfied.  Under the assumption that any volume of such mails would be
untenable in production, the chance of such configurations existing in
the wild seems even lower.  As such, the tiny cost of additional
Unprivileged users in such cases is worth the notable code
simplification.