fix ticket link name without rights on target ticket
If you have ticket 1 which is linked to ticket 2 and the user doesn't have
rights for ticket 2 the link is:
1: [] (owner ticket 2)
The ticket 2 subject and status isn't displayed because the rights check
but the ticket 2 owner is displayed without any rights check.

Also, if ticket 2 status is closed the class of the span element of the
link isn't set to 'ticket-inactive' because the ticket 2 status value
isn't accessible for the user.

With this commit the link name would fall back to $URI->AsString which shows
only the ticket id (the subject isn't shown in this case).