add $Type to %ARGS so we can specify the type of the textarea in the caller
by default, the type is controlled by the system config($MessageBoxRichText)
and user preferences. however, sometimes we want to explicitly set the type no
matter what system config or the user preference is.

this is initially for the IPs and Addresses textareas in RTIR's ScriptedAction
tool page(/RTIR/Tools/ScriptedAction.html), where rich texts are not useful at
all(actually harmful as the backend doesn't handle the rich format).