Split action handling into plugin classes
This allows mail plugins to define their own actions that they handle,
and dispatches handling an action to the first plugin that declares it
can handle it.  Using this, the handling of correspond and comment, as
well as the "unsafe" actions take and resolve, are moved out into their
own plugins.

As part of this, ParseCcAddressesFromHead, which was only called during
ticket creation, is inlined and placed in _HandleCreate in the new
RT::Interface::Email::Action::Defaults plugin, which is always enabled.

The $UnsafeEmailCommands configuration is now accomplished by adding
Action::Take and/or Action::Resolve to @MailPlugins.