Detect and offer to clean up links to missing articles
Article is added to @models to make it available in the sub that
checks for links to missing objects. However, @models is used
in several other checks in the validator and as a result missing
articles are now also detected in transactions and OCFVs.
Validator will offer to remove the transactions, but doesn't
yet handle OCFVs.