Allow users to set a "based on" category for external CFVs
The based on option was specifically not shown if the CF being
edited was flagged as an external CF, however it was still
possible to set it if you set the Field Value Source and the
based on option in the same edit. Regardless, categories appear
to work fine with external custom fields.