Use the correct -Values on select-multiple checkbox inputs
4b21442e changed to having a method to generate CF field names.
However, it incorrectly encoded that all "List" renderings (both radio
and checkboxes) should use the singular '-Value'.  Prior to 4b21442e,
the _only_ rendering which used '-Value' was the radio rendering, which
set $name explicitly; other locations hardcoded '-Values', or (in the case
of checkboxes) set $name to '-Values'.

This resulted in checkboxes using '-Value', which in turn caused the
same value to be be able to be re-added, and appear multiple times on
the same ticket; see [rt3 #29392].

Switch the GetCustomFieldInputName method to only return '-Value' for
the radio rendering of selects, restoring the same input names as prior
to 4b21442e.  It also removes a now-misleading comment, as the $name
variable is used for _all_ Select inputs, not just radio/checkboxes.  As
such, ensure that (in the absence of a $Name, which currently occurs
nowhere in core RT), the '-Value' option would still only be generated
for radiobuttons.