Parse old-style (not chained) Categrory drop-down arguments correctly
As of 4b21442, the string '-Category' is appended to the '-Value' or
'-Values', instead of replacing it.  Update all of the locations which
ignore the old-style Category to look for the suffix, instead of the
fixed string.

Without this, attempts to update custom field values with old-style
categories produce "User asked for an unknown update type for custom
field" warnings, and bulk update may erroneously set the CF values to
their selected category, rather than their selected value.