Make Crypt not an Auth:: plugin, but hardcoded
Always pushing Auth::Crypt onto plugins causes ordering difficulty -- in
that BeforeDecode plugins all see undecrypted content.  It is thus
imposssible for a plugin to to examine decrypted but not decoded
content.

Move Auth::Crypt, which actually does no auth, to no longer be a plugin,
but hardcoded into Gateway.