Warn about SetFieldsOnce
This callback was added in d77874ed to allow you to munge the %once
declared @fields variable.  If you used Default, you would keep
appending to the global shared variable.  When c07d62ad fixed the bug
that we were pushing Custom Fields onto the global shared variable by
switching from a %once to an %init, this callback became useless because
@fields was redeclared on every request.

The first commit was in 3.7.15, the second was in 3.7.86.  This means no
released version of RT has had a version of this callback which worked
properly.  We will remove it in 4.4 and add some dev notes here.

The correct solution is to use the Default callback.

CallbackOnce is still used properly with %once blocks in ColumnMap.