Ensure than an empty requestor is preserved across "Add More Files"
The use if $ARGS{Requestor} || 'something' as the default value for the
Requestor field means that if emptied, and the create does not go
through -- either because the "Add More Files" button was pressed, or
fields vailed validation, or an extension halted it -- then the
Requestor will pop back to being the current user.

Preserve the user's wishes, by using a defined-or to only set the
default if it was _unsubmitted_.  Note that this is a minor behavior
change, in that external links to Create.html?Requestors=&Queue=1
previously got the current user as the requestor, instead of no
requestors.  However, this is an unlikely scenario, and they are now
actually getting what they asked.

This leaves the remaining defaults to use ||; changing them is left to a
more comprehensive branch.