fixed failed tests caused by the new Entryhint change
in the cf admin page, the default EntryHint is bound to the type, but the
binding is in js, which is not easy to include to our tests. so we have to
manually set the default EntryHint in tests.

the reason we can't simply default EntryHint to it's FriendlyType is we allow
empty EntryHint, and it's not straightforward to distinguish undef from it in
submitted values of web form.