Remove ambiguity of "substitute ... for ..." in $CanonicalizeEmailAddressMatch
The phrase "substitute x for y" can be read to either mean "substitute
occurrences of x for y" (leaving y in the finished string) or
"substitute x for occurrences of y" (leaving x in the finished string).
Reword to use less ambiguous phrasing, and link more explicitly to to
CanonicalizeEmailAddress for overriding, in case of more complex
canonicalization.