When merging instances with identical $Organization, do not qualify names
Qualify() is meant to ensure that objects which are name-unique keep
their own "namespace" when merged across different RT instances.
Importing an RT into one with an identical $Organization is a fair sign
that they are from the same provenance, and thus should merge Groups and
Queues, instead of creating duplicates.