Save path because some installers or tests can change cwd
In cases where the CPAN installer fails to install GD and
its dependencies, the cwd can remain set to the directory
where GD was unpacked and tests were run like

  ~/.cpan/build/GDGraph-1.48-N0EX5K

When $0 is invoked later, rt-test-dependencies isn't found
because we're in the wrong directory.

The warning was masked in RT 4.0 because after the 'use'
to determine if Test::NoWarnings was installed, all warnings
were consumed by that module. This is fixed in RT 4.2 by ff15f21d3.