The lifecyles drop-down does not list approvals; prevent changing away from it
Pressing submit on the Approvals queue page set its lifecycle to
"default", as "approvals" is filtered out of the list of lifecycles to
display.  Once changed by accident, it is also impossible to undo.  If
changed intentionally, it may break lifecycles if the new lifecycle
contains different statuses or transitions.

Prevent this family of errors by hiding the lifecycle dropdown for the
approvals queue.