Hide the Disabled box on the Approvals queue
The Approvals queue is flagged by Diabled = 2; this presents as
"disabled" in the UI.  While marking it "enabled" does not cause many
problems, attempting to mark it re-Disabled leaves it as Disabled = 1,
which does not allow tickets to be created in it.

Hide the Disabled box entirely to prevent these sorts of problems.  This
also requires moving the SetEnabled hidden form field to only display if
Enabled was actually displayed.  The defaults for signing and encryption
are moved out of the SetEnabled block and inside their own SetCrypt
check, as their presence on the page is now orthogonal to the Enabled
checkbox.