tag SystemError transactions with an error class
This makes it easy to make the SystemError transactions more visible to the
user with a little bit of css.
Previously the txn was tagged with the 'other' class.
The new 'error' class use currently the same style as the previous 'other'
class to not surprise the user with an layout change within an stable release.