Default the content-transfer-encoding to 8bit
RT sets a content-transfer-encoding of 8bit on the top-level of outgoing
email, in RT::Interface::Email::SendEmail.  This sets the default for
the entire message.  Unfortunately, MIME::Entity is not aware of this,
and assumes that a lack of content-transfer-encoding header in any
particular part means it should default to 7bit.

These combine, in the case of signed email, to cause the "put a
content-transfer-encoding header" loop to mistakenly believe that a part
with 8bit data is actually 7bit, and thus it is left with no
content-transfer-encoding.

Assume that all unmarked parts are 8bit, as that is what the top-most
header implies anyway.  This causes RT to correctly apply a transfer
encoding to 8bit data in signed messages.