check both of "-Values" and "-Value" to find default values in ValidateCustomFields
in 4b21442, we refactored cf input name into a sub, which caused select cf
current values "not selected" when cloning a ticket. see also #29751.

turned out that we hardcoded "-Value" in the clone code before, which is not
consistent with the cf default name. 4b21442 reveals a potential bug that we
missed to check "-Values" in ValidateCustomFields.  this commit completes the
check(i.e. to check both "-Values" and "-Value"), like what we do for
$DECODED_ARGS in EditCustomField.