Prompt to password _after_ telling them how it will be sent
As written, if RTEXTERNALAUTH=1 was set, RT prompted for a password,
then told you if it was going to be sent in the clear, and immediately
sent it (while telling you that you could ^C it, but giving you no
opportunity to do so).

Reverse the order to match the non-externalauth case below, into the
useful order.