Allow GSSAPI authentication in bin/rt
dc653cb8, which introduced the optional GSSAPI authentication, did so in
a way that made it impossible to enable.  It required externalauth=undef
(and the modules be loadable) in order for $no_strong_auth to be unset,
which is required to skip the HTTP Basic and username/password auth
paths.  Unfortunately, it provided no way to _set_ an undef
externalauth, as the default value is set to 0 (despite the comment at
the top), and neither environment variables nor config file allow for
setting an undef value.

Rework the authentication paramters to bin/rt entirely, to clean this
up.  Specifically, inspect an "auth" parameter, which may be set to
"rt", "basic", or "gssapi" and reacts appropriately.  For backwards
compatibility, "externalauth=1" is equivalent to "auth=basic".  Choosing
GSSAPI authentication also explicitly requires the GSSAPI and
LWP::Authen::Negotiate modules, and aborts if they fail to be found,
rather than falling back to sending RT's built-in auth.

Fixes I#25074.