I#29595: Undef foolproof for PrimaryGroupBy on upgrade
In 4.0, when we process the submitted form for save chart operation
without changing any option in the dropdowns (either ChartStyle or
PrimaryGroupBy), then there isn't any value in the %ARGS which causes
these options to be saved as undef in DB. Additionally, that is the
same behavior for 4.2. Thus, we need to undef foolproof when handling
PrimaryGroupBy to avoid the warnings.