wrap with <pre> so HTML::FormatText::WithLinks::AndTables can recognize it
in 6b117af, we replaced <pre> with "<div style='white-space: pre-wrap; font-family: monospace;'>
so long lines can wrap when converting plain text to html.

this brings another problem when downgrading html to text back(this happens
when the email template is html): HTML::FormatText::WithLinks::AndTables
doesn't recognize the div, so the whitespaces in the plain text are not
preserved.

this commit wraps the "<div>" with "<pre>" so long lines can wrap and also
HTML::FormatText::WithLinks::AndTables can recognize and make the right
conversion.  (thanks to trs++)

see also #29185