Strip non-word characters from CF names for variable names
The existing Simple templates did not address the issue of CF names with
spaces in them.  This strips non-word characters to establish the
variable name -- a trasformation which might cause duplication of CF
names, but this does not notably worsen the situation, as they did not
already posess any uniqueness guarantees.

Note that this may still lead to unicode variable names, as \w is UTF-8
aware, allowing for the variable "$TicketCFStraße".

Fixes I#18446.