Force Disabled and Lifecycle on the ___Approvals queue
Because ___Approvals needs to be Disabled = 2 in order to be hidden but
still have tickets created, we fixed the display of the Enabled?
checkbox in f294fec4.  Since some users will have enabled and then
disabled the queue prior to coming up to 4.2, they will be unable to
create Approvals tickets without manually going and changing the value
in the database.

The same applies to the lifecycle; because we ceased listing the
"approvals" lifecycle in b97a61b6, merely pressing "Save Changes" on the
__Approvals queue would unfixably alter the lifecycle to "default"
instead of "approvals".

Force ___Approvals to have the Disabled and Lifecycle values assigned on
a new 4.2 install.