Only include closing paren if URL included an open paren
Open and close paren characters are valid in links.  However, links are
often placed in parenthetical statements, leading MakeClicky to
incorrectly include the closing paren in the link -- which, when
clicked, is thus often a 404.

Use a heuristic to determine if a last-character closing paren is likely
intended to be part of the link or not, based on if the link includes an
open paren character.  This will allow both of the following links to
still work as their writers likely intended:

     http://en.wikipedia.org/wiki/Fish_(disambiguation)
     (something like http://bestpractical.com/)

This is, of course, a trade-off.  However, the ability to link to
http://en.wikipedia.org/wiki/) is likely not a feature that will be
missed.

Fixes #29064.