Explicitly specify the plural, to help lead translators to do so
Translators may not be familiar with the %quant() form, and may simply
translate it as-is -- including only specifying the singular form.
Unless the language happens to pluralize by adding a trailing "s", this
will result in incorrect plural forms.

Help suggest that the language should specify both singular and plural
forms by doing so in English.  This does not change the English
pluralizations in any way, but instead serves to lead the translator to
do so in cases where it would be necessary.