Confirm loaded queue object before setting context object
For cases where an invalid Queue value is passed as an arg,
either a name or id of a non-existent queue or even 0, previously
the emtpy queue object would still be set as the context object
even though the load failed. This caused the search to add a
Limit on a LookupType of RT::Queue, which won't find any
custom fields.

Check the return code to avoid setting the empty queue object and
to warn so there is some help in the logs to track down a possible
typo in the queue name/id.