Decoding content, and returning characters, is incorrect
The whole purpose of this function is to return bytes, not characters.
Decoding the bytes to characters as utf8 defeats the purpose, and only
works because RT's internal representation of characters happens to be
utf8, which gets used by MIME::Entity to later encode the characters.

Switch to simply returning the bytes, as-is.