Allow for non-Queue context objects
This opens the door for using LoadByName for custom fields not on
Transactions, Tickets, or Queues.  Callers specify a LookupType and
ObjectId, from which the class to load (if not RT::Queue) can be
inferred.  Consequently, using LoadByName for User (or Asset) custom
fields no longer requires writing an alternate and parallel
implementation of LoadByName.