Allow callers to not get disabled CFs
For backwards compatibility, LoadByName has always maintained the
property that it may return a disabled custom field -- though it does so
only if there are no matching enabled fields.

Provide an option to control this behavior; the default is maintained as
it was previously, for backwards compatibility.

Note that previously, the only OrderBy was the Disabled flag -- this
means that if there were two non-disabled custom fields with the same
name, which one was returned was possibily non-deterministic.  This
commit places a predictable ordering on the resolution; namely, lower
SortOrder and lower id take precedence.