Simplify logic by loading the queue even if not strictly necessary
The existing logic in the "if" statement serves to skip the block in
precisely one situation -- the passed Queue is numeric, and there
already exists a ContextObject.  In such situation, it is not necessary
to attempt to load the Queue -- but it is not harmful, either.  In fact,
it may allow a better warning in cases where a numeric but non-existant
Queue is passed in.  It also clarifies the logic immeasurably.